pyproject.toml
[project]
name = "pygrader-utils"
version = "0.1.0"
description = "TBD"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "ipywidgets>=8.1.5",
  "mypy>=1.13.0",
  "numpy>=2.1.3",
  "pynacl>=1.5.0",
  "requests>=2.32.3",
  "ruff>=0.8.1",
  "types-requests>=2.32.0.20241016",
]
